Moved Geographical normalizers to subdirectory
