+++ src/Normalizer/FromJson/Geographical/GeographicalNameNormalizer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson;
+namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Geographical;
 
 use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
