+++ src/Normalizer/FromJson/Geographical/GeographicalNamesNormalizer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson;
+namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Geographical;
 
 use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
