+++ src/Normalizer/FromJson/Locality/LocalityNameNormalizer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Locality;
 
-use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\GeographicalNameNormalizer;
+use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Geographical\GeographicalNameNormalizer;
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\IdExtractor;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityNameId;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityName;
